+++ crates/cluster-client/zisk/src/lib.rs
@@ -12,5 +12,7 @@ mod error;
 #[cfg(test)]
 mod test;
 
-pub use client::ZiskClusterClient;
-pub use error::Error;
+pub use ere_prover_core::{Input, RemoteProverConfig};
+pub use ere_verifier_zisk::{ZiskProgramVk, ZiskProof};
+
+pub use crate::{client::ZiskClusterClient, error::Error};
