+++ crates/compiler/airbender/src/lib.rs
@@ -4,5 +4,5 @@ mod error;
 mod rust_rv32ima;
 
 pub use ere_compiler_core::*;
-pub use error::Error;
-pub use rust_rv32ima::AirbenderRustRv32ima;
+
+pub use crate::{error::Error, rust_rv32ima::AirbenderRustRv32ima};
