+++ crates/compiler/openvm/src/lib.rs
@@ -5,6 +5,8 @@ mod rust_rv32ima;
 mod rust_rv32ima_customized;
 
 pub use ere_compiler_core::*;
-pub use error::Error;
-pub use rust_rv32ima::OpenVMRustRv32ima;
-pub use rust_rv32ima_customized::OpenVMRustRv32imaCustomized;
+
+pub use crate::{
+    error::Error, rust_rv32ima::OpenVMRustRv32ima,
+    rust_rv32ima_customized::OpenVMRustRv32imaCustomized,
+};
