+++ crates/compiler/zisk/src/lib.rs
@@ -6,7 +6,8 @@ mod rust_rv64ima;
 mod rust_rv64ima_customized;
 
 pub use ere_compiler_core::*;
-pub use error::Error;
-pub use go_customized::ZiskGoCustomized;
-pub use rust_rv64ima::ZiskRustRv64ima;
-pub use rust_rv64ima_customized::ZiskRustRv64imaCustomized;
+
+pub use crate::{
+    error::Error, go_customized::ZiskGoCustomized, rust_rv64ima::ZiskRustRv64ima,
+    rust_rv64ima_customized::ZiskRustRv64imaCustomized,
+};
