+++ crates/prover/core/src/lib.rs
@@ -1,5 +1,4 @@
 #![cfg_attr(not(test), warn(unused_crate_dependencies))]
-#![allow(non_camel_case_types)]
 
 pub mod error;
 pub mod input;
@@ -9,8 +8,11 @@ pub mod resource;
 
 pub use ere_codec as codec;
 pub use ere_verifier_core::{PublicValues, zkVMVerifier};
-pub use error::CommonError;
-pub use input::Input;
-pub use prover::{ProgramVk, Proof, zkVMProver};
-pub use report::{ProgramExecutionReport, ProgramProvingReport};
-pub use resource::{ProverResource, ProverResourceKind, RemoteProverConfig};
+
+pub use crate::{
+    error::CommonError,
+    input::Input,
+    prover::{ProgramVk, Proof, zkVMProver},
+    report::{ProgramExecutionReport, ProgramProvingReport},
+    resource::{ProverResource, ProverResourceKind, RemoteProverConfig},
+};
