+++ crates/server/client/src/lib.rs
@@ -2,5 +2,6 @@
 
 mod client;
 
-pub use client::*;
 pub use ere_prover_core::*;
+
+pub use crate::client::*;
