+++ crates/util/compile/src/lib.rs
@@ -3,8 +3,10 @@
 mod error;
 mod rust;
 
-pub use error::CommonError;
-pub use rust::{
-    CargoBuildCmd, RustTarget, cargo_metadata, rustc_path, rustup_active_toolchain,
-    rustup_add_components, rustup_add_rust_src, rustup_add_target,
+pub use crate::{
+    error::CommonError,
+    rust::{
+        CargoBuildCmd, RustTarget, cargo_metadata, rustc_path, rustup_active_toolchain,
+        rustup_add_components, rustup_add_rust_src, rustup_add_target,
+    },
 };
