+++ crates/verifier/core/src/lib.rs
@@ -2,5 +2,5 @@ mod public_values;
 mod verifier;
 
 pub use ere_codec as codec;
-pub use public_values::PublicValues;
-pub use verifier::zkVMVerifier;
+
+pub use crate::{public_values::PublicValues, verifier::zkVMVerifier};
